chore: refactor collectible code out of packages/ui
