+++ apps/mobile/src/features/collectibles/components/collectible-gltf.tsx
@@ -2,8 +2,8 @@ import { ReactNode, useMemo } from 'react';
 import { StyleSheet } from 'react-native';
 import { WebView } from 'react-native-webview';
 
-import { CollectibleCard } from './collectible-card.native';
-import { CollectibleImage } from './collectible-image.native';
+import { CollectibleCard } from './collectible-card';
+import { CollectibleImage } from './collectible-image';
 
 interface CollectibleGltfProps {
   src: string;
