+++ apps/mobile/src/features/collectibles/components/collectible-html.tsx
@@ -2,9 +2,10 @@ import { ReactNode, useState } from 'react';
 import { StyleSheet } from 'react-native';
 import { WebView } from 'react-native-webview';
 
-import { Box, TouchableOpacity } from '../../../native';
-import { CollectibleCard } from './collectible-card.native';
-import { ImageUnavailable } from './image-unavailable.native';
+import { Box, TouchableOpacity } from '@leather.io/ui/native';
+
+import { CollectibleCard } from './collectible-card';
+import { ImageUnavailable } from './image-unavailable';
 
 interface CollectibleHtmlProps {
   src: string;
