+++ apps/mobile/src/features/collectibles/components/collectible-text.tsx
@@ -1,7 +1,8 @@
 import { WebView } from 'react-native-webview';
 
-import { Pressable, Text } from '../../../native';
-import { CollectibleCard } from './collectible-card.native';
+import { Pressable, Text } from '@leather.io/ui/native';
+
+import { CollectibleCard } from './collectible-card';
 
 interface CollectibleTextProps {
   src: string;
