+++ apps/mobile/src/features/collectibles/components/index.ts
@@ -1,0 +1,14 @@
+export { BnsImage } from './bns';
+export { CollectibleAudio } from './collectible-audio';
+export { CollectibleCard } from './collectible-card';
+export { CollectibleGltf } from './collectible-gltf';
+export { CollectibleHtml } from './collectible-html';
+export { CollectibleImage } from './collectible-image';
+export { CollectibleSvg } from './collectible-svg';
+export { CollectibleText } from './collectible-text';
+export { CollectibleTypeIconOverlay } from './collectible-type-icon-overlay';
+export { CollectibleVideo } from './collectible-video';
+export { ImageUnavailable } from './image-unavailable';
+export { Inscription } from './inscription';
+export { Sip9 } from './sip9';
+
