+++ apps/mobile/src/features/token/bitcoin/stamp.tsx
@@ -2,7 +2,7 @@ import { TokenDetailsProps } from '@/features/token/types';
 import { imageUnavailableLabel } from '@/features/token/utils/image-unavailable-label';
 
 import { StampAsset } from '@leather.io/models';
-import { CollectibleImage, ImageUnavailable } from '@leather.io/ui/native';
+import { CollectibleImage, ImageUnavailable } from '@/features/collectibles/components';
 import { getAssetId, serializeAssetId } from '@leather.io/utils';
 
 interface StampProps {
