+++ apps/mobile/src/features/token/collectibles-list.tsx
@@ -10,7 +10,7 @@ import { EmptyCollectiblesState } from '@/features/token/components/empty-collec
 import { TokenDetailsProps } from '@/features/token/types';
 
 import { CollectibleView } from '@leather.io/features';
-import { CollectibleTypeIconOverlay } from '@leather.io/ui/native';
+import { CollectibleTypeIconOverlay } from '@/features/collectibles/components';
 import { assertUnreachable } from '@leather.io/utils';
 
 import { Inscription } from './bitcoin/inscription';
