+++ apps/mobile/src/queries/collectibles/account-collectibles.query.ts
@@ -5,7 +5,7 @@ import { useCollectiblesFlag } from '@/features/feature-flags';
 import { useAccountAddresses } from '@/hooks/use-account-addresses';
 import { useQuery } from '@tanstack/react-query';
 
-import { createCollectibleView } from '@leather.io/features';
+import { type CollectibleView, createCollectibleView } from '@leather.io/features';
 import { AccountAddresses, CryptoAssetId } from '@leather.io/models';
 import { createAccountCollectiblesQueryConfig } from '@leather.io/queries';
 import { SerializedCryptoAssetId, deserializeAssetId, matchesAssetId } from '@leather.io/utils';
@@ -17,12 +17,14 @@ export function useAccountCollectibleByAssetId(
 ) {
   const account = useAccountAddresses(fingerprint, accountIndex);
 
-  return toFetchState(useAccountCollectibleByAssetIdQuery(account, deserializeAssetId(assetId)));
+  return toFetchState<CollectibleView[]>(
+    useAccountCollectibleByAssetIdQuery(account, deserializeAssetId(assetId))
+  );
 }
 
 export function useAccountCollectibles(fingerprint: string, accountIndex: number) {
   const account = useAccountAddresses(fingerprint, accountIndex);
-  return toFetchState(useAccountCollectiblesQuery(account));
+  return toFetchState<CollectibleView[]>(useAccountCollectiblesQuery(account));
 }
 
 function useSanitizedAccount(account: AccountAddresses) {
