+++ packages/features/package.json
@@ -3,6 +3,7 @@
   "version": "1.1.2",
   "license": "MIT",
   "type": "module",
+  "types": "./dist/index.d.ts",
   "scripts": {
     "build": "tsdown",
     "build:watch": "tsdown --watch",
