--- packages/ui/src/components/collectibles/index.native.ts
@@ -1,10 +1,0 @@
-export * from './native/collectible-html.native';
-export * from './native/collectible-image.native';
-export * from './native/collectible-svg.native';
-export * from './native/collectible-text.native';
-export * from './native/collectible-type-icon-overlay.native';
-export * from './native/inscription.native';
-export * from './native/collectible-card.native';
-export * from './native/image-unavailable.native';
-export * from './native/bns.native';
-export * from './native/sip9.native';
