--- packages/ui/src/components/collectibles/index.web.ts
@@ -1,8 +1,0 @@
-export * from './web/collectible-audio.web';
-export * from './web/collectible-image.web';
-export * from './web/collectible-text.web';
-export * from './web/collectible-other.web';
-export * from './web/collectible-iframe.web';
-export * from './web/image-unavailable.web';
-export * from './web/inscription-preview-card/index.web';
-export * from './web/collectible-item.layout.web';
