--- packages/ui/src/components/collectibles/web/collectible-iframe.web.tsx
@@ -1,34 +1,0 @@
-import { ReactNode, useState } from 'react';
-
-import { CollectibleItemLayout, CollectibleItemLayoutProps } from './collectible-item.layout.web';
-import { Iframe } from './iframe.web';
-import { ImageUnavailable } from './image-unavailable.web';
-
-interface CollectibleIframeProps extends Omit<CollectibleItemLayoutProps, 'children'> {
-  icon: ReactNode;
-  src: string;
-}
-export function CollectibleIframe({ icon, src, ...props }: CollectibleIframeProps) {
-  const [isError, setIsError] = useState(false);
-
-  if (isError)
-    return (
-      <CollectibleItemLayout collectibleTypeIcon={icon} {...props}>
-        <ImageUnavailable />
-      </CollectibleItemLayout>
-    );
-
-  return (
-    <CollectibleItemLayout collectibleTypeIcon={icon} {...props}>
-      <Iframe
-        aspectRatio="1 / 1"
-        height="100%"
-        objectFit="cover"
-        onError={() => setIsError(true)}
-        src={src}
-        width="100%"
-        zIndex={99}
-      />
-    </CollectibleItemLayout>
-  );
-}
