--- packages/ui/src/components/collectibles/web/image-unavailable.web.tsx
@@ -1,15 +1,0 @@
-import { styled } from 'leather-styles/jsx';
-
-import { Eye1ClosedIcon } from '../../../icons/eye-1-closed-icon.web';
-import { CollectiblePlaceholderLayout } from './collectible-placeholder.layout.web';
-
-export function ImageUnavailable() {
-  return (
-    <CollectiblePlaceholderLayout>
-      <Eye1ClosedIcon />
-      <styled.span pt="space.02" px="space.04" textStyle="label.03">
-        Image currently unavailable
-      </styled.span>
-    </CollectiblePlaceholderLayout>
-  );
-}
