--- packages/ui/src/components/collectibles/web/inscription-preview-card/components/inscription-metadata.web.tsx
@@ -1,33 +1,0 @@
-import { ReactElement } from 'react';
-
-import { Flex, styled } from 'leather-styles/jsx';
-
-import { Link } from '../../../../link/link.web';
-
-interface InscriptionMetadataProps {
-  action?(): void;
-  actionLabel?: string;
-  icon?: ReactElement;
-  subtitle: string;
-  title: string;
-}
-export function InscriptionMetadata({
-  action,
-  actionLabel,
-  icon,
-  subtitle,
-  title,
-}: InscriptionMetadataProps) {
-  return (
-    <Flex alignItems="flex-start" flexDirection="column" justifyContent="center">
-      {icon && icon}
-      <styled.span textStyle="label.01">{title}</styled.span>
-      <styled.span textStyle="caption.01">{subtitle}</styled.span>
-      {action ? (
-        <Link color="stacks" onClick={() => action()} textStyle="caption.01" variant="text">
-          {actionLabel}
-        </Link>
-      ) : null}
-    </Flex>
-  );
-}
