--- packages/ui/src/components/collectibles/web/inscription-preview-card/index.web.ts
@@ -1,4 +1,0 @@
-export { InscriptionPreviewCard } from './inscription-preview-card.web';
-export { InscriptionTextLayout } from './components/inscription-text.layout.web';
-export { InscriptionPreviewLayout } from './components/inscription-preview.layout.web';
-export { InscriptionImage } from './components/inscription-image.web';
